# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "201 Project API",
        "version": "1.0"
    },
    "paths": {
        "/api/groups/list": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "List all the groups and their IDs",
                "responses": {
                    "200": {
                        "description": "Groups have been retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "properties": {
                                                "id": {
                                                    "type": "int"
                                                },
                                                "name": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": [
                                        {
                                            "id": 1,
                                            "name": "201"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/groups/items": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get a group's subscibed items",
                "operationId": "App\\Http\\Controllers\\GroupController::items",
                "responses": {
                    "200": {
                        "description": "Group items have been retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "properties": {
                                                "id": {
                                                    "type": "int"
                                                },
                                                "group_id": {
                                                    "type": "int"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "pic_url": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": [
                                        {
                                            "id": 1,
                                            "group_id": 1,
                                            "name": "Banana",
                                            "in_stock": 1,
                                            "pic_url": null
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/groups/subscribe/{group_id}": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Subscribe a user to a group",
                "operationId": "App\\Http\\Controllers\\GroupController::subscribe",
                "responses": {
                    "200": {
                        "description": "User has been subscribed to group",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "User has been subscribed to group 201"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/api/groups/create": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Create a new group and subscribe to it",
                "operationId": "App\\Http\\Controllers\\GroupController::create",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Group created and user subscribed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": {
                                        "id": "1",
                                        "name": "201"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
...
        "/api/items/subscribe/{item_id}": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Subscribe a user to an Item",
                "operationId": "App\\Http\\Controllers\\ItemController::subscribe",
                "responses": {
                    "200": {
                        "description": "User subscribed to Item",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": {
                                        "id": "1",
                                        "group_id": "1",
                                        "name": "Banana",
                                        "in_stock": "1"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
...
    },
    "components": {
        "responses": {
            "401": {
                "description": "Request is not authorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "error": "Unauthorized"
                        }
                    }
                }
            },
            "400": {
                "description": "Bad request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "error": "Bad request"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT"
            }
        }
    }
}